Everyone gets added to a global team
Add a team: value to the -conf.yml file. If it exists then everyone
will get added to that team.

If a user has no team in the .csv file, it is OK. That user will get
added to the team: in the -conf.yaml file (if any). Program won't crash.